backend/requirements/requirements.txt
#
# This file is autogenerated by pip-compile with python 3.10
# To update, run:
#
#    pip-compile --output-file=requirements/requirements.txt requirements/requirements.in
#
amqp==5.1.1
    # via kombu
asgiref==3.5.2
    # via django
async-timeout==4.0.2
    # via redis
billiard==3.6.4.0
    # via celery
celery[redis]==5.2.7
    # via
    #   -r requirements/requirements.in
    #   django-celery-beat
    #   django-celery-results
certifi==2022.12.7
    # via
    #   requests
    #   sentry-sdk
cffi==1.15.1
    # via cryptography
charset-normalizer==2.1.1
    # via requests
click==8.1.3
    # via
    #   celery
    #   click-didyoumean
    #   click-plugins
    #   click-repl
click-didyoumean==0.3.0
    # via celery
click-plugins==1.1.1
    # via celery
click-repl==0.2.0
    # via celery
cryptography==38.0.4
    # via pyjwt
defusedxml==0.7.1
    # via python3-openid
dj-rest-auth[with_social]==2.2.5
    # via -r requirements/requirements.in
django==4.1.4
    # via
    #   -r requirements/requirements.in
    #   dj-rest-auth
    #   django-allauth
    #   django-celery-beat
    #   django-debug-toolbar
    #   django-extensions
    #   django-filter
    #   django-guardian
    #   django-hashid-field
    #   djangorestframework
    #   djangorestframework-simplejwt
django-allauth==0.50.0
    # via
    #   -r requirements/requirements.in
    #   dj-rest-auth
django-celery-beat==2.4.0
    # via -r requirements/requirements.in
django-celery-results==2.4.0
    # via -r requirements/requirements.in
django-debug-toolbar==3.8.1
    # via -r requirements/requirements.in
django-extensions==3.2.1
    # via -r requirements/requirements.in
django-filter==22.1
    # via -r requirements/requirements.in
django-grappelli==3.0.4
    # via -r requirements/requirements.in
django-guardian==2.4.0
    # via -r requirements/requirements.in
django-hashid-field==3.3.7
    # via -r requirements/requirements.in
django-simple-history==3.2.0
    # via -r requirements/requirements.in
django-softdelete==0.9.4
    # via -r requirements/requirements.in
django-timezone-field==5.0
    # via django-celery-beat
djangorestframework==3.14.0
    # via
    #   -r requirements/requirements.in
    #   dj-rest-auth
    #   djangorestframework-simplejwt
djangorestframework-simplejwt==5.2.2
    # via -r requirements/requirements.in
gunicorn==20.1.0
    # via -r requirements/requirements.in
hashids==1.3.1
    # via django-hashid-field
hiredis==2.1.0
    # via redis
idna==3.4
    # via requests
kombu==5.2.4
    # via celery
markdown==3.4.1
    # via -r requirements/requirements.in
oauthlib==3.2.2
    # via requests-oauthlib
prompt-toolkit==3.0.36
    # via click-repl
psycopg2-binary==2.9.5
    # via -r requirements/requirements.in
pycparser==2.21
    # via cffi
pygments==2.13.0
    # via -r requirements/requirements.in
pyjwt[crypto]==2.6.0
    # via
    #   django-allauth
    #   djangorestframework-simplejwt
python-crontab==2.7.0
    # via django-celery-beat
python-dateutil==2.8.2
    # via python-crontab
python3-openid==3.2.0
    # via django-allauth
pytz==2022.7
    # via
    #   celery
    #   django-timezone-field
    #   djangorestframework
pyyaml==6.0
    # via -r requirements/requirements.in
redis[hiredis]==4.4.0
    # via
    #   -r requirements/requirements.in
    #   celery
requests==2.28.1
    # via
    #   django-allauth
    #   requests-oauthlib
requests-oauthlib==1.3.1
    # via django-allauth
sentry-sdk==1.12.1
    # via -r requirements/requirements.in
six==1.16.0
    # via
    #   click-repl
    #   django-softdelete
    #   python-dateutil
sqlparse==0.4.3
    # via
    #   django
    #   django-debug-toolbar
tzdata==2022.7
    # via django-celery-beat
uritemplate==4.1.1
    # via -r requirements/requirements.in
urllib3==1.26.13
    # via
    #   requests
    #   sentry-sdk
vine==5.0.0
    # via
    #   amqp
    #   celery
    #   kombu
wcwidth==0.2.5
    # via prompt-toolkit
wheel==0.38.4
    # via django-softdelete
whitenoise==6.2.0
    # via -r requirements/requirements.in

# The following packages are considered to be unsafe in a requirements file:
# setuptools
